+++ backend/analytics-service/src/main/java/the/chak/ecommerce/analytics/repository/CategoryAggregate.java
@@ -1,8 +1,10 @@
 package the.chak.ecommerce.analytics.repository;
 
+import java.math.BigDecimal;
+
 /**
  * Revenue for one category. {@code categoryId} is null for sales whose product has no known
  * category -- the caller groups those under a single "uncategorized" entry.
  */
-public record CategoryAggregate(Long categoryId, String categoryLabel, double revenue) {
+public record CategoryAggregate(Long categoryId, String categoryLabel, BigDecimal revenue) {
 }
